User Update API auth.js moved into userSlice.js.
