+++ frontend/src/store/auth/index.js
@@ -8,45 +8,6 @@ const initialState = {
   error: null,
 };
 
-// Profile Change API
-export const changeProfile = createAsyncThunk(
-  "auth/changeProfile",
-  async (data, { rejectWithValue }) => {
-    try {
-      const response = await axios({
-        method: "PUT",
-        url: "/api/v1/private/user/",
-        data: data,
-        headers: {
-          "Content-Type": "application/json",
-        },
-      });
-      return response.data;
-    } catch (error) {
-      return rejectWithValue(error.response.data.message || error.message);
-    }
-  }
-);
-
-// Change Password API
-export const changePassword = createAsyncThunk(
-  "changePassword",
-  async (data, { rejectWithValue }) => {
-    try {
-      const response = await axios({
-        method: "PUT",
-        url: "/api/v1/private/user/password",
-        data: data,
-        headers: {
-          "Content-Type": "application/json",
-        },
-      });
-      return response.data;
-    } catch (error) {
-      return rejectWithValue(error.response.data.message || error.message);
-    }
-  }
-);
 
 const authSlice = createSlice({
   name: "auth",
@@ -54,40 +15,7 @@ const authSlice = createSlice({
   reducers: {},
   extraReducers: (builder) => {
     builder
-      .addCase(changeProfile.pending, (state) => {
-        state.loading = true;
-        state.error = null;
-      })
-      .addCase(changeProfile.fulfilled, (state, action) => {
-        state.data = action.payload;
-        state.loading = false;
-        showToast("dismiss");
-        showToast("success", state.data.message);
-      })
-      .addCase(changeProfile.rejected, (state, action) => {
-        state.loading = false;
-        state.error = action.payload;
-        showToast("dismiss");
-        showToast("error", state.error);
-      });
-
-    builder
-      .addCase(changePassword.pending, (state) => {
-        state.loading = true;
-        state.error = null;
-      })
-      .addCase(changePassword.fulfilled, (state, action) => {
-        state.data = action.payload;
-        state.loading = false;
-        showToast("dismiss");
-        showToast("success", state.data.message);
-      })
-      .addCase(changePassword.rejected, (state, action) => {
-        state.loading = false;
-        state.error = action.payload;
-        showToast("dismiss");
-        showToast("error", state.error);
-      });
+      
   },
 });
