+++ frontend/src/store/user/userSlice.js
@@ -1,6 +1,7 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import { createSlice } from "@reduxjs/toolkit";
 import axios from "axios";
+import { showToast } from "src/utils/showToast";
 
 const initialState = { //initial state
   loading: false,
@@ -25,6 +26,47 @@ export const fetchProfileUser = createAsyncThunk(
   }
 );
 
+// Profile Change API
+export const changeProfile = createAsyncThunk(
+  "auth/changeProfile",
+  async (data, { rejectWithValue }) => {
+    try {
+      const response = await axios({
+        method: "PUT",
+        url: "/api/v1/private/user/",
+        data: data,
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
+      return response.data;
+    } catch (error) {
+      return rejectWithValue(error.response.data.message || error.message);
+    }
+  }
+);
+
+// Change Password API
+export const changePassword = createAsyncThunk(
+  "changePassword",
+  async (data, { rejectWithValue }) => {
+    try {
+      const response = await axios({
+        method: "PUT",
+        url: "/api/v1/private/user/password",
+        data: data,
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
+      return response.data;
+    } catch (error) {
+      return rejectWithValue(error.response.data.message || error.message);
+    }
+  }
+);
+
+
 const userSlice = createSlice({
   name: "user",
   initialState: initialState,
@@ -42,6 +84,38 @@ const userSlice = createSlice({
       state.loading = false; //set loading
       state.error = action.payload; //set error to payload
     }) 
+    .addCase(changeProfile.pending, (state) => {
+      state.loading = true;
+      state.error = null;
+    })
+    .addCase(changeProfile.fulfilled, (state, action) => {
+      state.data = action.payload;
+      state.loading = false;
+      showToast("dismiss");
+      showToast("success", state.data.message);
+    })
+    .addCase(changeProfile.rejected, (state, action) => {
+      state.loading = false;
+      state.error = action.payload;
+      showToast("dismiss");
+      showToast("error", state.error);
+    })
+    .addCase(changePassword.pending, (state) => {
+      state.loading = true;
+      state.error = null;
+    })
+    .addCase(changePassword.fulfilled, (state, action) => {
+      state.data = action.payload;
+      state.loading = false;
+      showToast("dismiss");
+      showToast("success", state.data.message);
+    })
+    .addCase(changePassword.rejected, (state, action) => {
+      state.loading = false;
+      state.error = action.payload;
+      showToast("dismiss");
+      showToast("error", state.error);
+    });
     }
 });
